Shortening subject if it's too large
If the forwarded message has too long subject line (hard do say how long is "too long", probably >256 characters), script canoot forward that message and returns an error.
Now, we short the subject to 100 characters and add "(...)" at the end of subject line. In order not to lose the content of subject, it is added to message body at the very top of it.